requirements.txt

fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.3


SQLAlchemy==2.0.43
psycopg2-binary==2.9.10

alembic==1.16.5
greenlet==3.2.4


redis==6.4.0


qdrant-client==1.15.1
grpcio==1.74.0
protobuf==6.32.1


openai==1.107.2
numpy==2.3.3
tqdm==4.67.1


python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1


python-multipart==0.0.20


python-dotenv==1.1.1
requests==2.31.0
PyYAML==6.0.2
click==8.2.1
colorama==0.4.6
typing_extensions==4.15.0


PyPDF2==3.0.1
python-docx==1.2.0
lxml==6.0.1
Mako==1.3.10
MarkupSafe==3.0.2
certifi==2025.8.3
urllib3==2.5.0
watchfiles==1.1.0
websockets==15.0.1
email-validator==2.2.0
tenacity==9.1.2
